you can make the call and see the results, assuming that you have shipalot running as a localhost:4000!

+++ db/migrate/20140226180451_adding_weight_and_shipping_info.rb
@@ -1,6 +1,6 @@
 class AddingWeightAndShippingInfo < ActiveRecord::Migration
   def change
-    add_column :products, :weight, :float
+    add_column :products, :weight, :integer
     add_column :products, :height, :integer
     add_column :products, :length, :integer
     add_column :products, :depth, :integer
